Make the Router abstract class.
